+++ operator/apis/http/v1alpha1/zz_generated.deepcopy.go
@@ -161,6 +161,11 @@ func (in *HTTPScaledObjectSpec) DeepCopyInto(out *HTTPScaledObjectSpec) {
 		*out = new(int32)
 		**out = **in
 	}
+	if in.PollingInterval != nil {
+		in, out := &in.PollingInterval, &out.PollingInterval
+		*out = new(int32)
+		**out = **in
+	}
 	if in.ScalingMetric != nil {
 		in, out := &in.ScalingMetric, &out.ScalingMetric
 		*out = new(ScalingMetricSpec)