+++ CHANGELOG.md
@@ -26,6 +26,7 @@ This changelog keeps track of work items that have been completed and are ready 
 
 - **General**: TODO ([#TODO](https://github.com/kedacore/http-add-on/issues/TODO))
 - **General**: Fix kubectl active printcolumn ([#1211](https://github.com/kedacore/http-add-on/issues/1211))
+- **General**: Allow configuration of PollingInterval in HTTPScaledObject ([#1251](https://github.com/kedacore/http-add-on/issues/1251))
 
 ### Improvements